Update HasJsonRelationships.php
Allow the array of objects to be the value of json field itself

options: is json the field in the examples below

These are the 3 scenarios that will cover all the bases. 

1.) options: [1,2,3] | options->role_ids

2.) options: {"user_id": 5, "roles": [ {"role_id": 1, "active": false}, {"role_id": 2, "active": true} ] }  | options->roles[]->role_id

*NEW*
3.)  options: [ {"role_id": 1, "active": false}, {"role_id": 2, "active": true}, {"role_id": 3, "active": false} ] | options[]->role_id 

There are times where user stores data like in the 3.) example. 

For 3.) the relationship would look like this ...  return $this->belongsToJson('App\Role', 'options[]->role_id');

I tested with attach(), detach(), sync() and toggle() and everything seems to work per normal. Would be a great improvement as this would expand the use case to these three which are the only possible scenarios.